--- assets/src/components/editing/models/image/resizer/Resizer.tsx
@@ -1,26 +1,0 @@
-// let resizer;
-//   if (ReactEditor.isFocused(editor) && selected && imageRef.current) {
-//     resizer = (
-//       <>
-//         <div
-//           style={Object.assign(boundResizeStyles('border'), {
-//             position: 'absolute',
-//             border: '2px solid rgb(0, 150, 253)',
-//             backgroundColor: 'rgba(0, 0, 0, 0)',
-//             zIndex: 30,
-//             borderColor: 'rgb(26, 115, 232)',
-//           })}
-//         ></div>
-//         {resizeHandle('nw')}
-//         {resizeHandle('n')}
-//         {resizeHandle('ne')}
-//         {resizeHandle('w')}
-//         {resizeHandle('e')}
-//         {resizeHandle('sw')}
-//         {resizeHandle('s')}
-//         {resizeHandle('se')}
-//       </>
-//     );
-//   } else {
-//     resizer = null;
-//   }
