+++ assets/styles/authoring/model-editors.scss
@@ -161,14 +161,6 @@
 
   .image-editor {
     clear: both;
-
-    .image-resize {
-      margin-left: auto;
-      margin-right: auto;
-      overflow: hidden;
-      resize: both;
-      height: auto;
-    }
   }
 
   .settings-input {
